v1.2.1 release

+++ src/Nacos/V2/Config/Impl/FileLocalConfigInfoProcessor.cs
@@ -188,7 +188,7 @@
                 {
                     try
                     {
-                        file.Delete();
+                        if (file.Exists) file.Delete();
                     }
                     catch
                     {
